tests: add CheckTest + Organização dos testes
