+++ LagChessApplication/Domains/Chess/ChessHistory.cs
@@ -5,6 +5,7 @@
         private readonly Stack<ChessMove> _moves = new();
 
         public ChessMove[] Get() => [.. _moves];
+        public ChessMove LastMove => _moves.First();
 
         public void Add(ChessMove move) => _moves.Push(move);
         public void Undo() => _moves.Pop();
